+++ Eby_AssignmentComplete.py
@@ -1,4 +1,4 @@
-import ebybrownlibserver as libserver
+import Eby_MessageProcessor as libserver
 import GlobalConstants
 import Eby_Message
 import re # regular expressions
